requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
branca==0.8.1
cachetools==5.5.2
census==0.8.24
CensusData==1.15.post1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.2
cssselect==1.3.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
feedfinder2==0.0.4
feedparser==6.0.11
ffmpy==0.5.0
filelock==3.18.0
folium==0.19.6
fonttools==4.58.0
fredapi==0.5.2
frozenlist==1.6.0
fsspec==2025.5.0
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
gradio==5.30.0
gradio_client==1.10.1
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
lightning-utilities==0.14.3
llvmlite==0.44.0
lxml==5.4.0
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
multidict==6.4.4
narwhals==1.40.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
newsapi-python==0.2.7
newspaper3k==0.2.8
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pandoc==2.4
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pipreqs==0.5.0
platformdirs==4.3.8
plotly==6.1.1
plumbum==1.9.0
ply==3.11
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pyngrok==7.2.8
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
pyzillow==0.7.0
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
rpds-py==0.25.1
ruff==0.11.10
safehttpx==0.1.6
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sgmllib3k==1.0.0
shap==0.47.2
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
sodapy==2.2.0
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.1
streamlit_folium==0.25.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tinycss2==1.4.0
tinysegmenter==0.3
tldextract==5.3.0
toml==0.10.2
tomlkit==0.13.2
torch==2.2.2
torchaudio==2.2.2
torchmetrics==1.7.1
torchvision==0.17.2
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
wordcloud==1.9.4
xyzservices==2025.4.0
yarg==0.1.9
yarl==1.20.0
